Changed n steps to uint8

+++ src/sonic_protocol/protocols/protocol_v1_0_0/procedure_commands/procedure_fields.py
@@ -37,7 +37,7 @@ field_type_f_shift = FieldType(
 )
 
 field_type_n_steps = FieldType(
-    field_type=np.uint32,
+    field_type=np.uint8,
     min_value=DeviceParamConstantType.MIN_N_STEPS
 )
